Give scans control over cache via scan dispatchers #1383
This commit enables per scan control over cache usages.  This was done
via extending the scope of ScanDispatchers to include cache control.

The built in SimpleScanDispatcher was improved to support mapping
scanner execution hints of the form scan_type=X to
cache usage directives for a scan.

Accumulo caches open files.  Before this change the cache was bound to
the file when it was opened.  Now a cache can be bound to an already
open file. The internal interface CacheProvider was created to
facilitate this.