fix #1408 remedy the ITs broken by #1389
Most of the broken ITs had issues with the change from dir uri to dir
name.  However the problem with VolumeChooserFailureIT was different, it
expected the created table Fate op to call the volume chooser.  The
volume chooser is no longer called in create table master code after
called until a tablet was assigned to a tserver.

Some of the failure behavior in VolumeChooserFailureIT is covered by the
PreferredVolumeChooserTest unit test.  However the behavior of ensuring
that when a volume chooser fails that the system does not fall back to
something else is not.   I could not think of a way to ensure the
fallback behavior does not happen, so I deleted the test. The lack of
any fallback behavior was just causing the create table call to hang
forever (it used to throw an exception).  I don't think we want to check
that a create table calls hangs for some minimum time.