Made tablet assignment wait on upgrade. (#1405)
While working on #1389 I noticed that tablets were being assigned to tservers
before their metadata was upgraded.  This commit makes assignment wait until a
tablets metadata is upgraded.  For example, assignment of the root tablet should
wait for zookeeper to be upgraded,  assignment of metadata tablets should wait
for the root tablet to be upgraded, and assignment of user tablets should wait
for metadata tablets to be upgraded.

The way waiting is done is by considering the upgrade status when computing
a tablets goal state.  If a tablets metadata store is not upgraded, then its
goal state will be unassigned.