Streamline passing scan parameters within tserver (#1437)
While working on #1383 I found it difficult to get caching parameters
from one part of the tserver to another for a scan.  The difficulty was
caused by many methods with many parameters that were mostly the same.
This commit changes the code to pass a single object around.

I decided to make these changes separate from #1383 inorder to make it
easier to review the changes for #1383.