Improve Upgrader9to10 code (#1441)
This changes the batching strategy for the Upgrader9to10 class.

The previous strategy was to fill a data structure until JVM memory was
sufficiently used up (50% of heap size), then process the upgrade for
those entries, and repeat until no more entries are left to upgrade.

This new strategy batches candidates for upgrade in batches of size 4
million characters (approx. 8MB batches), regardless of memory.

This stabilizes the testing, as it is much simpler to reproduce and test
fixed-size batches, than it is to manipulate the JVM heap size during
testing. As a result, many improvements to the GCUpgrade9to10TestIT were
made here. This also fixes that IT, which began failing with the use of
the G1GC instead of CMS (#1427), because G1GC didn't work well with a
master configured with a 16MB JVM heap and crashed frequently OOMEs.