Add code to replace relative paths on upgrade. #1397 (#1461)
* Resolve and replace all relative tablet file paths found in metadata tables with absolute
paths during upgrade. Absolute paths are resolved by prefixing relative paths
with a volume configured by the user in a the instance.volumes.upgrade.relative
property, which is only used during an upgrade. If any relative paths are found and this property
is not configured, or if any resolved absolute path does not correspond to a
file that actually exists, the upgrade step fails and aborts without making changes.