Fix #1350 Update ListVolumesUsed to support root table metadata
With changes in #1313 the root tablet file metadata is no different from
any other tablet. Given that fact, ListVolumesUsed has been updated to
remove the specialized code that it had for zookeeper and instead make
use of the Ample API for interfacing with zookeeper. Additionally
removed the use of Scanner from ListTable.