Backport gc metrics to 1.9.x to use hadoop metrics2. (#1406)
* Backport gc metrics to 1.9.x to use hadoop metrics2.

This backport bring the gc metrics from master / 2.0 into the 1.9.x branch. The metrics
can be enabled / disabled with GC_METRICS_ENABLED property (default false for 1.9.x)  The
metrics changes include:

- Publish current gc metrics to hadoop 2 metrics reporting system.
- Add gc run counter that increments on each gc cycle run
- Add metric to track time required for gc post op (compact, flush, none)

Also Includes:

- Move IT resource file for PR #1381 to correct directory so test passes
in CI and on the command-line. Previously, it may have only worked in
some IDEs, because of the way those IDEs put test resources on the class
path.

* checkpoint: warning clean-up

* Code clean-up and prep for review

one open todo - check if metadata consistency check can be added as a metric.

* defer master check reproting

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>