Remove miniDFS from HalfDeadTServerIT
Remove use of miniDFS from IT, because sometimes miniDFS is slow to
start up, and it causes the IT to time out. It is not needed for the
test, since the test case works fine on LocalFileSystem-based DFS.

Use java.util.scanner to clean up boilerplate BufferedReader stuffs.

Suppress some spammy ZK logs in IT log4j2 config.

Add prefix to DumpOutput class in HalfDeadTServerIT to help troubleshoot
the IT in future.