Remove miniDFS from HalfDeadTServerIT (#1545)
Remove use of miniDFS from IT, because sometimes miniDFS is slow to
start up, and it causes the IT to time out. It is not needed for the
test, since the test case works fine on LocalFileSystem-based DFS.

Use java.util.scanner to clean up boilerplate BufferedReader stuffs.

Suppress some spammy ZK logs in IT log4j2 config.

Add prefix to DumpOutput class in HalfDeadTServerIT to help troubleshoot
the IT in future.

Build shared library for fake_disk_failure.so only once, before the
class.

Use ProcessBuilder to connect the IO when building the shared library,
so any errors from the compiler are shown on the IT's console capture.

Remove compilation warning about unused variable in fake_disk_failure.c

Add comments and make some variables a bit more explicit.