Use sets instead of arrays for volumes
Originally ACCUMULO-3376, this change replaces the use of `String[]`
with `Set<String>` throughout the VolumeChooser and VolumeManager code.
This makes the API for VolumeChooser a bit more friendly, making it
easier to use Streams and lambdas to filter and transform volumes
throughout the code.

This change also adds some more sanity checks on the instance volumes
configuration (checking for empty list and duplicates).

This also includes a utility to warn about planned API changes in
internal pluggable code (such as the VolumeChooser), but without being
too spammy.

Update the default `VolumeChooser.choosable()` to return the full set of
volumes, so that all are checked for flush and sync support on tserver
startup. This choosable mechanism is a bit dubious anyway, because
choosers could make different decisions over time (in response to
configuration changes, for example) than they do at startup. So,
defaulting to checking for the maximal set seems like a better strategy
out-of-the-box.

Make ServerConstants.baseUris unmodifiable.