Remove some reflection where not needed (#1550)
Remove unnecessary reflection to load methods that are contained in
public stable Hadoop APIs.

This reflection was added for simultaneous Hadoop 1 and Hadoop 2
support. We now require Hadoop 3 and these APIs are stable, so
reflection is not needed.

Also reduce visibility of some methods in TTimeoutTransport exposed for
testing and delete unnecessary code.