Use instanceof to check for ViewFs
The mechanism deployed in Hadoop's ViewFileSystemUtil to check if a
FileSystem is ViewFs does so by checking the URI scheme. However that
code is not stable public API, so I had previously implemented the log
closer to check for ViewFs using the same implementation, but using only
stable public APIs from Hadoop. That didn't work, because RawFileSystem
doesn't implement the getScheme method. So, use instanceof instead, even
though ViewFileSystem itself is an evolving API (even though it is
evolving, the class name is unlikely to change).