Revert "Update Last Location. Solution for #1169 (#1453)"
This reverts commit d12ff6c700e003202fd2e9842499324561246b15.

The reverted commit was not ready for merging, as it caused dozens of
test failures (around 60), and appeared to fundamentally break parsing
of many stored locations by causing locations to be parsed with
HostAndPort.fromString(location) that were not created with
HostAndPort.toString(). This may be a more fundamental problem with some
of the code in Ample; further investigation and testing is needed before
merging.