Remove negations in 'isEmpty' and 'if' statements.
This is a follow-up to #1590. After updating code to use isEmpty() this
ticket re-factors to switch order of statements to remove the negation.
Additionally 'if' statement negations were re-factored. A couple of cases
where a statement contained double negations were updated also. The
intent is to increase the clarity of the statements.

This update does not alter the behavior of the code. Hopefully it just
results in more readable code. Since this is primarily a stylistic
update if there is opposition to it then it can be ignored and the
ticket can be closed.