Lambdas replaced with method references. (#1602)
This ticket updates instances of lambdas which can be replaced with method references. There was also one instance of an anonymous type being replaced with a method reference.

One instance was not changed due to possible subtle differences in the way in which constructor method references can behave. See the notes from PR #1602 if interested in details.