Removed a duplicate 'if' conditionin TabletLocatorImpl.
The 'updateCache' method within TabletLocatorImpl.java contains two
identical 'if' conditions, both of which throw 'IllegalStateExceptions'.
I removed the second, given that it would never be reached if the
condition is true. Even though the conditional checks are identical, the
message provided to the exceptions differ. I kept the first since that
would be the one presented to the user if this condition was satisfied.