Fix FateConcurrencyIT periodic failures
This patch fixes periodic FateConcurrencyIT failures due to insufficient
threads available for concurrent compactions, caused by the introduction
of the property
'tserver.compaction.major.service.default.planner.opts.executors' to
replace 'tserver.compaction.major.concurrent.max'.

This change now sets a sufficient number of executors to run the test
consistenly, when one tserver tries to compact 3 or 4 of the 4 total
tables, and the other is responsible for compacting only 0 or 1.

Additionally:

* Fix a javadoc issue for the improper use of less-than symbol in
  DefaultCompactionPlanner
* Use a non-deprecated property as an example in MiniAccumuloRunner
  instead of the deprecated 'tserver.compaction.major.concurrent.max'
* Set instance-wide compaction property only once instead of each time a
  SlowOps object is constructed
* Inline unneeded private SlowOps constructor
* Fix typo in FateConcurrencyIT javadoc
* Use Streams to simplify code logic
* Avoid creating unused SlowOps in `@Before` method in FateConcurrencyIT
* Clean up use of TimeUnits and conversions (using nano in an additional
  place, and variables which are named to indicate they are millis)