Add libraries in tool.sh script for 1.10 examples
When attempting to run the 1.10.x Accumulo examples with Hadoop version
3.x, there were several NoClassDefFoundError exceptions thrown when
using the tool.sh script to run MR jobs. This PR updates the tool.sh
script to add the necessary jar files to the lib path allowing the MR
jobs to run successfully. The additions work with both Hadoop 2.x and
Hadoop 3.x.