Add tablename option to shell insert command. (#1771)
This change adds the --table/-t option to the insert command when inside the shell or when using the accumulo command with 'accumulo shell -e <task>'. The option is not required, but if used it allows data to be inserted into a table in situations where the shell is not within a table context or when in a table context to insert data into another table.

Although I would not expect this option to be utilized very often, I found myself in need of it when working with the various accumulo examples in the accumulo-examples repo. I had many situations where I wished to insert some test data from a bash shell using the 'accumulo shell -e insert r f q v'  form but this would fail due to there being no table context for the insertion. This update allows data inserts from a command line using the 'accumulo shell' format.