Check for consistency of table-related shell commands.
This ticket examined the table-related shell commands to examine if they
provided a means to indicate the table name. The 'insert' command was
updated recently to add the ability to supply a target table. This
ticket provides the same capability for the 'delete' command.

In all, there were 39 shell commands that relate to table operations. At
this point they should all allow a target table name to be supplied.