Fixes #1808 - Stop server side VM on Error in Thread
Modified the AccumuloUncaughtExceptionHandler to log Exception and
Error and Halt the VM if the system property HaltVMOnThreadError
is set to true. Modified all Executors to use the NamingThreadFactory
which uses AccumuloUncaughtExceptionHandler. When modifying
SimpleTimer I found that some critical tasks could fail and only
be logged. Created SimpleCriticalTimer for the critical tasks.