Update SystemCredentials to use crypt(3) format
This fixes #1810

This updates the SystemToken generation to use SHA-512 while hashing the
system's `instance.*` configuration properties (including
`instance.secret`) by using commons-codec's crypt(3) implementation.

It also adds more javadocs and makes the code a bit more readable, while
removing a hard-coded global constant and an unneeded new instance
property introduced in #1798. That property is unneeded, since the
existence of the property itself creates wire-incompatibility. Instead,
this change explicitly bumps the wire version.

Also add sanity checks in SystemCredentialsTest for the wire version
values and to detect significant changes in the default algorithm used
by commons-codec to hash the credentials to prevent suprises when
upgrading the commons-codec library in future.