Fix a bug in the AccumuloClient builder
This fixes #1839

Fix a bug in the ClientContext's from(Properties) method, which is used
to build a new AccumuloClient from an existing set of properties. This
is often used to initialize a builder, and then to subsequently modify
the client via the builder to use different credentials, or other
settings, that were stored in the provided Properties.

This fix modifies the from(Properties) method to make a copy of the
provided Properties, so that subsequent builder activity does not modify
the Properties object that the caller provided.