Fix #1792 flaky/broken GcMetricsIT
Rewrite portions of GcMetricsIT to improve its reliability and to add
comments and rename methods to make it more clear how it is intended to
work, so it's easier to debug in future.

Also use daemon threads for the tailer, and ensure it is started
consistently (using a dedicated method) and closed properly when done.

Remove unnecessary catching of exceptions, and allow them to be thrown
from the tests instead, because they will be handled by the test
framework.