Fixes #1808: refactor creation Threads/ThreadPools, set UncaughtExceptionHandler on all threads
Fixes #1808: refactor creation Threads/ThreadPools, set UncaughtExceptionHandler on all threads

Created a single code path for Threads, Executors, Timers, etc. All threads are now named, are
daemon threads, and have the AccumuloUncaughtExceptionHandler set on them. The
AccumuloUncaughtExceptionHandler logs when an unhandled Exception is thrown and prints to
sterr and calls Halt when an Error is thrown.