Rename master packages to manager.
This change renames all packages with 'master' in the name to 'manager'
with the exception of the o.a.a.core.master.thrift package. Due to the
usage of some of these classes in the user-facing TabletBalancer code
hierarchy, the renaming of this package will wait until the next major
release.

re #1642