Rename master packages to manager. (#1904)
Rename master packages to manager.

This change renames all packages with 'master' in the name to 'manager'
with the exception of the o.a.a.core.master.thrift package. Due to the
usage of some of these classes in the user-facing TabletBalancer code
hierarchy, the renaming of this package will wait until the next major
release. While it would have been possible to rename the package for
most of the classes int he master.thrift package, then we would have
been left in a state where we could not re-generate the thrift classes
if that were needed.

The TableState enum, which is in a renamed package, is used in some
classes in the public API though it is not used in the public API
itself (i.e., it is not a parameter or return type for any public or
protected method in the public API). The usages are limited to internal
comparisons and lookup of enum values based on the string. Therefore,
this renaming won't break any clients.

* IDE refactoring was used to rename packages having master in their
  name to the same package name, but with master replaced by manager.
* Updated Property enums where the default implementation for the
  property was a class that got re-packaged. Kept a "compatibility"
  implementation for such classes, which is the class under the old
  package name extends the class under the new package name, is
  deprecated, is empty except for a constructor that logs a warning
  about the deprecation.
* Add "compatibility" classes for those in packages renamed that have
  a main method. The compatibility class extends the class in the new
  package name, and the main method delegates to the moved class' main
  method as well as logging a warning about the deprecation.

re #1642