Fix log file names in minicluster and two broken ITs. (#1914)
Commit e6164901a3acdd464a7a9beff0126097b53fc06f removed the main method
from Manager. In order to work around this, minicluster was modified to
use KeywordExecutable classes to launch. However, this caused all
minicluster log files to be named "Main" since the main class is always
org.apache.accumulo.start.Main. This could be worked around with a
log4j2 properties file and variable name to cause the file logger to use
the right class name, but then that would require all minicluster users
to use the supplied logger configuration, or replicate the same
configuration. Instead, the main method was added back to Manager so
that minicluster logging will work as expected. This also fixes the
broken ITs BackupManagerIT and ThriftServerBindsBeforeZooKeeperLockIT
which were both broken since they were attempting to reference the main
method in the Manager class.