External Compactions feature (#1451)
Added a feature to perform major compactions outside of the TabletServer. This
feature is optional and disabled by default. External compactions can be
accomplished by using the type 'external' in the compaction service configuration
for a table and starting two new server processes, the CompactionCoordinator and
the Compactor. External compactions:

  * Allow compactions to outlive a TabletServer
  * Allow compactions to occur concurrent to a Tablet being re-hosted
  * Reduce the load on the TabletServer, giving it more cycles to insert mutations and respond to scans
  * Allow compactions to scale up and down independent of tablet server scaling and tablet locations. Tablet location does not constrain where compactions can run.  This allows tablets on a few tservers with lots of compaction work to spread the work out in ways not possible before.


Co-authored-by: Keith Turner <kturner@apache.org>