Unit test converting client props to BatchWriterConfig
Created two unit tests for the conversion of client properties to
BatchWriterConfig. One using default values and another using modified
values.

Refactored getBatchWriterConfig as suggested in initial ticket.

Added code in getBatchWriterConfig to allow for updating
BATCH_WRTIER_THREADS_MAX as it was missing in the method as it existed.