Document some cache props as requiring restart (#2164)
Add certain blockcache-related properties to the "fixed" properties
list, which will ensure they are documented as not taking effect until
after a restart. This is already the current behavior, because the cache
manager is instantiated as a singleton only at tserver startup, but this
helps ensure that is predictably the case, and will document it as
requiring a restart in our property documentation.

Also organize the fixed properties into logical groups, just to make it
easier to maintain slightly.