Update trace code in GarbageCollectionAlgorithm (#2216)
Update location of TraceScope code in GarbageCollectionAlgorithm. With the GCA code now using an iterator rather tnan continue point, the seperate call to getCandidates with GCA is no longer needed as its only purpose purpose was to allow the GCE method to be traced. That method was removed. The trace was instead placed around the call to readCandidatesThatFitInMemory as that is the method responsible for collecting and batching up deletion candidates.