Versioned Properties - refactored to address PR comments (#2224)
Adds VersionedProperties to manage related properties as a single entity.

Versioned properties are intended to be store and managed properties as a group (system, namespace, table) stored on a single ZooKeeper node rather than node / individual properties. The versioning is to allow efficient checking of cached entries vs stored entries.
  - Encoding / Decoding with optional compression
  - Test example that allows for encryption

This is the first step in replacing property management from single properties (in ZooKeeper) to storing related
properties on a single ZooKeeper node.