Add compaction coordinator and compactor to cluster start / stop scripts
Modified the cluster start/stop scripts to use a new file (cluster.yaml)
for defining the hosts that will run the different server components. Added
a class (and test) that parses the yaml file into a form that is usable by
the scripts. Added ability to specify and start/stop the external compaction
server processes.

Closes #2138


Co-authored-by: Keith Turner <kturner@apache.org>