Fix incorrect scan range output in getsplits command
In the Accumulo shell, calling getsplits with the verbose option can
result in incorret output. It occurs when the tableId of the table
happens to be a single character and there are other tables where the
tableId starts with the same character. This results in the output of
getsplits displaying splits for the other tables as well.

Closes #2356