Fix coordinator check in Monitor (#2392)
* Changes for the monitor external compactions page to detect 
when the Coordinator can't be found or has moved. The page itself will 
throw pop-up errors with a message that is not very helpful. With this 
change you can still see those errors but after a minute the monitor will 
detect that its gone and show the user friendly message on the page.
* Also fix log message in ECResource