Add trivial ZK Watcher to MiniAccumuloClusterImpl (#2434)
Add a trivial watcher to MiniAccumuloClusterImpl.verifyUp to avoid an
annoying ERROR coming from ZooKeeper that is logged when a ZooKeeper
client is constructed with a null Watcher.

Also address comments on #2414 to remove the redundant use of
requireNonNull for calls to Optional.get(). Instead, use a private
helper method to wait for the Process to be recorded as having a start
time, up to 10 seconds. So, instead of expecting the start time to be
non-null, it is now expected to be non-null within 10 seconds, and it
does so using Optional.isEmpty() instead of requireNonNull.

This fixes #2431