Remove self-assignment statement. (#2445)
Removed a self-assignment of the hostPortString variable to itself.

If the hostPortString is non-null, nothing happens and the variable is unchanged. If the hostPortString is null, a NullPointerException is thrown. There is no need to assign it to itself.

Change suggested by errorprone analysis.

Added a message to the requireNonNull method indicating the cause of the NullPointerException.

Also, added java.util.Objects import remove need for fully qualified path to requireNonNull.