Remove redundant native map loading
Remove redundant call to NativeMapLoader.load() in NativeMap, which
isn't needed because the TabletServerResourceManager already loads it
when the tserver starts up. This static initializer caused problems with
test references to NativeMap, and tried to load the native map library
before the test had a chance to load it from the search path.

Instead, if NativeMap is attempted to be used without loading the native
library, it will just naturally throw an UnsatisfiedLinkError.

This fixes #2450