Remove unnecessary TServerUtils stop method
With newer versions of Thrift, the executor service used inside the
server is reachable by calling `getInvoker()`. So, it is no longer
necessary to use reflection to try to access the private field to shut
it down. This change removes that reflection code and shuts down the
executor service in the stop() method of our custom subclass. Since that
is substantially simpler, it is no longer necessary to have a dedicated
static utility method to stop a server. That method is removed and
stop() is called directly.