Swap in reload4j instead of log4j1.2
For the 1.10.x branch, use reload4j instead of the EOL log4j1.2, which
contains several CVEs and is no longer supported by its Apache project.
Reload4j is the continuation of log4j1.2, forked from the Apache
project by the original log4j1 creator. It is largely a drop-in
replacement for log4j1.2.