Added FaTE TStatus.SUBMITTED (#2462)
When a FaTE transaction is created in the Manager its status is set
to IN_PROGRESS even though it's not actually running, it could be
waiting to run. The output from the `fate print` command will show
that the newly created transaction is IN_PROGRESS and could be
confusing to users. This change introduces the SUBMITTED state, which
is the state of the transaction after it has been created but before
it is executed.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>