Added mechanism to cancel FaTE operations from Shell with Manager up
This change adds a new Thrift API method to the FateServiceHandler to
cancel an existing FaTE transaction. The API method is called from
FateCommand using the ManagerClient class. This will attempt to set
the status to failed for a non-running transaction and will interrupt
the TransactionRunner thread for a transaction that is currently in
progress