Enable ScanServer to load multiple and different versions of a Tablet.
This was accomplished by using a subclass of KeyExtent in the OnlineTablets
object so that the key in the map was a different object, even if the extent
information was the same. ScanServerKeyExtent subclasses KeyExtent, has a
UUID member variable, and overrides compareTo, hashCode, and equals. Added a
new IT, ScanServerConcurrentTabletScanIT, which loads data into a table and
creates scanners at different times and checks that each scanner only sees the
data that was loaded at the time of scanner creation.