Remove load-jars-hdfs from bin/accumulo-util
Remove the jar loading utility, which is broken, from the accumulo-util
script. This utility was intended as a convenience to help users copy
jars to HDFS to be used with the VFS classloader. However, that utility
is in process of being separated from Accumulo's core, and so this tool
makes a lot of assumptions about the user's classloader configuration.
Furthermore, it's not really necessary for Accumulo to maintain a helper
tool for copying files for the user, as they can copy files on their
own, according to their specific needs.

This fixes #2440