Add errorpone profile to main Accumulo pom. (#2447)
Added a new profile, errorprone, to the pom.xml file. Use of the profile
allows the Google ErrorProne static analysis tool to examine the
Accumulo codebase and flag errors/warnings.

See https://errorprone.info for additional info and
https://errorprone.info/bugpatterns for a list of bug patterns available.

Users withing to use the tool can examine the bug pattern list and add
items of interest specifically to the profile. Any patterns or warnings that
should always be flagged can then be permanently added to the profile
as needed.

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>
Co-authored-by: Dom G. <domgarguilo@apache.org>