Some cleanup of Zoo-related code
Remove three unneeded ZooCache constructors - these were simple
variations of the main one, with very little added value, and removing
them made some method calls slightly less convenient, but enabled a lot
of code complexity reduction.

Remove unneeded ZooConfigurationFactory - it was unnecessarily complex,
written to handle caching of ZooConfiguration objects for multiple
InstanceIds; however, it was only ever used by
ServerConfigurationFactory, which only ever called it once, with the
first instanceId it found. This is sufficient. If further work is needed
to ensure ServerConfigurationFactory only uses a single InstanceId, that
work should be done in that class. This change preserves the existing
behavior, but much more simplified. An additional benefit to removing
this is that it can no longer be confused with the similarly named
ZooCacheFactory.

Remove unneeded "hack" to use a watcher in ZooCacheFactory that supplied
an unused Watcher to force the ZooCacheFactory to return a new instance.
Now, it explicitly returns a new instance, with a method that says as
much in its name and javadoc. The old method wasn't used for anything
else.

Remove unnecessary call in ReplicationIT to ZooCacheFactory, instead
using the ZooCache that is already available in the ClientContext in the
test.