Remove log4j2 config file from accumuo-test jar
This change removes the log4j2-test.properties file from the
accumulo-test.jar, but keeps it on the class path for the unit and
integration tests in the test module. This ensures our tests log useful
information during the build testing, but avoids leaking that log4j
configuration into the class path in other situations where the
accumulo-test.jar might be used for other kinds of testing.

This fixes #2527

Also move the shellit.shellit file to complete #2497 and remove the last
item from the root of our jars.

Also tweak the system property configuration of maven-failsafe-plugin in
the root pom.xml to ensure the configuration there is applied even when
the test module sets its own additional system properties for the
plugin.