Fate op defer time is not used
A comment was added to Fate.java to indicate how the deferTime variable
is being used. It clarifies that the value is used in the manner of an
exit code for the particular if/else loop after the comment.

But it also makes clear that the value of deferTime is used as a wait
value later on within the ZooStore class.

No code was modified with this ticket.