Check scheduled tasks periodically for errors (#2524)
Updated code that use ScheduleExecutorServices and were not retaining a reference to the ScheduleFuture object returned by the methods that create scheduled tasks. If the code was adding a non-scheduled task via the `submit` method, I changed the code to use the `execute` method. For code that was adding a scheduled task, but not doing anything with the ScheduledFuture, I modified the code to pass the ScheduledFuture reference to a utility method in ThreadPools. The new ThreadPools utility methods periodically check the ScheduledFuture to see if it has completed and if it has thrown an error. If an error was thrown from a critical scheduled task, then an `Error` is thrown from the ThreadPools task checking thread which will be caught by the AccumuloUncaughtExceptionHandler, which will log the Error and terminate the VM.